# Chore: improve performance of `indent` rule