--- a/lib/rules/indent.js
+++ b/lib/rules/indent.js
@@ -150,12 +150,12 @@ class BinarySearchTree {
     }
 
     /**
-     * Finds the entry with the largest key greater than or equal to the provided key
+     * Finds the entry with the largest key greater than the provided key
      * @param {number} key The provided key
      * @returns {{key: number, value: *}|null} The found entry, or null if no such entry exists
      */
-    findGe(key) {
-        const iterator = this._rbTree.ge(key);
+    findGt(key) {
+        const iterator = this._rbTree.gt(key);
 
         return iterator && { key: iterator.key, value: iterator.value };
     }
@@ -385,11 +385,11 @@ class OffsetStorage {
         const fromTokenDescriptor = fromTokenIsInRange && this._getOffsetDescriptor(fromToken);
 
         // First, remove any existing nodes in the range from the tree.
-        let existingNodeInRange = this._tree.findGe(range[0]);
+        let existingNodeInRange = this._tree.findGt(range[0]);
 
         while (existingNodeInRange && existingNodeInRange.key < range[1]) {
             this._tree.delete(existingNodeInRange.key);
-            existingNodeInRange = this._tree.findGe(range[0]);
+            existingNodeInRange = this._tree.findGt(range[0]);
         }
 
         // Insert a new node into the tree for this range