--- a/lib/rules/indent.js
+++ b/lib/rules/indent.js
@@ -150,23 +150,22 @@ class BinarySearchTree {
     }
 
     /**
-     * Finds the entry with the largest key greater than the provided key
-     * @param {number} key The provided key
-     * @returns {{key: number, value: *}|null} The found entry, or null if no such entry exists
-     */
-    findGt(key) {
-        const iterator = this._rbTree.gt(key);
-
-        return iterator && { key: iterator.key, value: iterator.value };
-    }
-
-    /**
-     * Removes an entry from the tree, if it exists
-     * @param {number} key The key of the entry
+     * Deletes all of the keys in the interval [start, end)
+     * @param {number} start The start of the range
+     * @param {number} end The end of the range
      * @returns {void}
      */
-    delete(key) {
-        this._rbTree = this._rbTree.remove(key);
+    deleteRange(start, end) {
+        const iterator = this._rbTree.ge(start);
+        const keysToDelete = [];
+
+        while (iterator.valid && iterator.key < end) {
+            keysToDelete.push(iterator.key);
+            iterator.next();
+        }
+        keysToDelete.forEach(key => {
+            this._rbTree = this._rbTree.remove(key);
+        });
     }
 }
 
@@ -385,12 +384,7 @@ class OffsetStorage {
         const fromTokenDescriptor = fromTokenIsInRange && this._getOffsetDescriptor(fromToken);
 
         // First, remove any existing nodes in the range from the tree.
-        let existingNodeInRange = this._tree.findGt(range[0]);
-
-        while (existingNodeInRange && existingNodeInRange.key < range[1]) {
-            this._tree.delete(existingNodeInRange.key);
-            existingNodeInRange = this._tree.findGt(range[0]);
-        }
+        this._tree.deleteRange(range[0] + 1, range[1]);
 
         // Insert a new node into the tree for this range
         this._tree.insert(range[0], descriptorToInsert);