--- a/lib/rules/indent.js
+++ b/lib/rules/indent.js
@@ -156,6 +156,11 @@ class BinarySearchTree {
      * @returns {void}
      */
     deleteRange(start, end) {
+
+        // Exit without traversing the tree if the range has zero size.
+        if (start === end) {
+            return;
+        }
         const iterator = this._rbTree.ge(start);
         const keysToDelete = [];