--- a/lib/rules/indent.js
+++ b/lib/rules/indent.js
@@ -162,15 +162,11 @@ class BinarySearchTree {
             return;
         }
         const iterator = this._rbTree.ge(start);
-        const keysToDelete = [];
 
         while (iterator.valid && iterator.key < end) {
-            keysToDelete.push(iterator.key);
+            this._rbTree = this._rbTree.remove(iterator.key);
             iterator.next();
         }
-        keysToDelete.forEach(key => {
-            this._rbTree = this._rbTree.remove(key);
-        });
     }
 }