--- a/lib/rules/indent.js
+++ b/lib/rules/indent.js
@@ -118,6 +118,8 @@ const KNOWN_NODES = new Set([
 
 /**
  * A mutable balanced binary search tree that stores (key, value) pairs. The keys are numeric, and must be unique.
+ * This is intended to be a generic wrapper around a balanced binary search tree library, so that the underlying implementation
+ * can easily be swapped out.
  */
 class BinarySearchTree {