--- a/lib/rules/indent.js
+++ b/lib/rules/indent.js
@@ -137,7 +137,13 @@ class BinarySearchTree {
      * @returns {void}
      */
     insert(key, value) {
-        this._rbTree = this._rbTree.remove(key).insert(key, value);
+        const iterator = this._rbTree.find(key);
+
+        if (iterator.valid) {
+            this._rbTree = iterator.update(value);
+        } else {
+            this._rbTree = this._rbTree.insert(key, value);
+        }
     }
 
     /**