<!--
    ESLint adheres to the [JS Foundation Code of Conduct](https://js.foundation/community/code-of-conduct).
-->

**What is the purpose of this pull request? (put an "X" next to item)**

[x] Other, please explain:

<!--
    If the item you've checked above has a template, please paste the template questions below and answer them. (If this pull request is addressing an issue, you can just paste a link to the issue here instead.)
-->

<!--
    Please ensure your pull request is ready:

    - Read the pull request guide (http://eslint.org/docs/developer-guide/contributing/pull-requests)
    - Include tests for this change
    - Update documentation for this change (if appropriate)
-->

<!--
    The following is required for all pull requests:
-->

**What changes did you make? (Give an overview)**

This updates the `indent` rule to be more performant by storing information on ranges of tokens rather than performing operations on many individual tokens one-at-a-time.

This results in a *19% end-to-end performance increase (!)* according to our `npm run perf` benchmark.

<details>
<summary>End-to-end benchmark</summary>

`master` branch:

```
Single File:
  CPU Speed is 3100 with multiplier 13000000
  Performance Run #1:  6601.645335ms
  Performance Run #2:  6537.859522ms
  Performance Run #3:  6469.447266ms
  Performance Run #4:  6777.504465ms
  Performance Run #5:  6431.605835ms

  Performance budget exceeded: 6537.859522ms (limit: 4193.548387096775ms)
```

`indent-performance` branch:

```
Single File:
  CPU Speed is 3100 with multiplier 13000000
  Performance Run #1:  5287.6544109999995ms
  Performance Run #2:  5284.148783ms
  Performance Run #3:  5326.879172ms
  Performance Run #4:  5338.387146ms
  Performance Run #5:  5340.651033ms

  Performance budget exceeded: 5326.879172ms (limit: 4193.548387096775ms)
```
</details>

<details>
<summary>Rule-specific benchmark on ESLint codebase</summary>

`master` branch:

```
$ TIMING=1 eslint lib/ conf/ bin/ tests/ --no-eslintrc --rule 'indent: [error, 4, {SwitchCase: 1}]' --env=es6 --no-inline-config
Rule   | Time (ms) | Relative
:------|----------:|--------:
indent |  5448.737 |   100.0%
```

`indent-performance` branch:

```
$ TIMING=1 eslint lib/ conf/ bin/ tests/ --no-eslintrc --rule 'indent: [error, 4, {SwitchCase: 1}]' --env=es6 --no-inline-config
Rule   | Time (ms) | Relative
:------|----------:|--------:
indent |  4610.891 |   100.0%
```
</details>


**Is there anything you'd like reviewers to focus on?**

I tried to write descriptive comments to explain what's going on. Is there anything that's unclear or could be improved?