@not-an-aardvark 2017-07-12
I've made a few small improvements -- now this improves end-to-end performance by 20%.

<details>
<summary>Benchmark</summary>

`master`:

```
Single File:
  CPU Speed is 3100 with multiplier 13000000
  Performance Run #1:  5877.433702ms
  Performance Run #2:  5805.863323ms
  Performance Run #3:  5738.451947ms
  Performance Run #4:  5712.380049ms
  Performance Run #5:  5754.480289ms

  Performance budget exceeded: 5754.480289ms (limit: 4193.548387096775ms)
```

`indent-performance`:

```
Single File:
  CPU Speed is 3100 with multiplier 13000000
  Performance Run #1:  4577.530063ms
  Performance Run #2:  4641.629504ms
  Performance Run #3:  4603.666983ms
  Performance Run #4:  4684.530936ms
  Performance Run #5:  4614.500893ms

  Performance budget exceeded: 4614.500893ms (limit: 4193.548387096775ms)
```

(I think my laptop has a different battery level than when I ran the benchmark in the original PR, so these two runs are comparable to each other, but they're not comparable to the original two runs in terms of absolute time.)
</details>